fix: 파이어베이스 초기화되는 버그 수정 (#44)
