+++ src/firebase/sampleData2.json
@@ -1,0 +1,124 @@
+{
+  "users": {
+    "user1": {
+      "username": "음악애호가",
+      "email": "music_lover@example.com",
+      "profileImg": "https://example.com/profiles/user1.jpg",
+      "playlistCount": 3,
+      "totalLikes": 50,
+      "totalForks": 5
+    },
+    "user2": {
+      "username": "케이팝팬",
+      "email": "kpop_fan@example.com",
+      "profileImg": "https://example.com/profiles/user2.jpg",
+      "playlistCount": 2,
+      "totalLikes": 30,
+      "totalForks": 2
+    }
+  },
+  "playlists": {
+    "playlist1": {
+      "userId": "user1",
+      "title": "신나는 K-POP 모음",
+      "description": "활기찬 하루를 위한 K-POP 플레이리스트",
+      "category": "K-POP",
+      "createdAt": "2024-08-15T09:00:00Z",
+      "updatedAt": "2024-08-16T10:30:00Z",
+      "likeCount": 25,
+      "forkCount": 3,
+      "commentCount": 5,
+      "videoCount": 3,
+      "thumbnailUrl": "https://example.com/thumbnails/playlist1.jpg",
+      "isPublic": true,
+      "videoIds": ["video1", "video2", "video3"]
+    },
+    "playlist2": {
+      "userId": "user2",
+      "title": "잔잔한 발라드",
+      "description": "마음을 위로하는 발라드 모음",
+      "category": "Ballad",
+      "createdAt": "2024-08-14T15:00:00Z",
+      "updatedAt": "2024-08-14T15:00:00Z",
+      "likeCount": 15,
+      "forkCount": 1,
+      "commentCount": 2,
+      "videoCount": 2,
+      "thumbnailUrl": "https://example.com/thumbnails/playlist2.jpg",
+      "isPublic": true,
+      "videoIds": ["video4", "video5"]
+    }
+  },
+  "videos": {
+    "video1": {
+      "youtubeId": "dQw4w9WgXcQ",
+      "title": "BTS - Dynamite",
+      "thumbnailUrl": "https://img.youtube.com/vi/dQw4w9WgXcQ/0.jpg",
+      "duration": 199
+    },
+    "video2": {
+      "youtubeId": "Cb70gcTVvYI",
+      "title": "BLACKPINK - How You Like That",
+      "thumbnailUrl": "https://img.youtube.com/vi/Cb70gcTVvYI/0.jpg",
+      "duration": 181
+    },
+    "video3": {
+      "youtubeId": "mPVDGOVjRQ0",
+      "title": "TWICE - Fancy",
+      "thumbnailUrl": "https://img.youtube.com/vi/mPVDGOVjRQ0/0.jpg",
+      "duration": 213
+    },
+    "video4": {
+      "youtubeId": "eN5mG_yMDiM",
+      "title": "IU - Through the Night",
+      "thumbnailUrl": "https://img.youtube.com/vi/eN5mG_yMDiM/0.jpg",
+      "duration": 259
+    },
+    "video5": {
+      "youtubeId": "3gyP8Di1qho",
+      "title": "Paul Kim - Every Day, Every Moment",
+      "thumbnailUrl": "https://img.youtube.com/vi/3gyP8Di1qho/0.jpg",
+      "duration": 241
+    }
+  },
+  "userPlaylists": {
+    "user1": {
+      "created": ["playlist1"],
+      "forked": [],
+      "liked": ["playlist2"]
+    },
+    "user2": {
+      "created": ["playlist2"],
+      "forked": ["playlist1"],
+      "liked": ["playlist1"]
+    }
+  },
+  "comments": {
+    "playlist1": {
+      "comment1": {
+        "userId": "user2",
+        "username": "케이팝팬",
+        "content": "정말 신나는 플레이리스트네요! 제 플레이리스트에도 추가해볼게요.",
+        "createdAt": "2024-08-15T14:30:00Z",
+        "updatedAt": "2024-08-15T14:30:00Z"
+      },
+      "comment2": {
+        "userId": "user1",
+        "username": "음악애호가",
+        "content": "좋아해 주셔서 감사합니다! 더 좋은 곡들 추가할게요.",
+        "createdAt": "2024-08-15T15:00:00Z",
+        "updatedAt": "2024-08-15T15:00:00Z"
+      }
+    }
+  },
+  "playlistVideos": {
+    "playlist1": {
+      "videoIds": ["video1", "video2", "video3"],
+      "order": [0, 1, 2]
+    },
+    "playlist2": {
+      "videoIds": ["video4", "video5"],
+      "order": [0, 1]
+    }
+  }
+}
