+++ src/pages/Search.tsx
@@ -1,3 +1,6 @@
-const Search = () => <div>Search</div>;
+const Search = () => {
+  const a = 2;
+  return <div>Search</div>;
+};
 
 export default Search;
